[PT-5] network separate local provider
