+++ data/lib/providers/local/dish_data_provider.dart
@@ -1,0 +1,6 @@
+import '../../entity/dish_entity/dish_entity.dart';
+
+abstract class DishDataProvider {
+  Future<void> saveAllDishes(List<DishEntity> input);
+  List<DishEntity> fetchAllDishes();
+}
