+++ data/lib/providers/local/theme_data_provider.dart
@@ -1,0 +1,6 @@
+abstract class ThemeDataProvider{
+  Future<void> saveTheme(bool input);
+  Future<void> saveScaleFactor(double input);
+  double fetchScaleFactor();
+  bool fetchTheme();
+}
