+++ data/lib/providers/local/user_data_provider.dart
@@ -1,0 +1,5 @@
+abstract class UserDataProvider{
+  Future<void> saveUser(String input);
+  bool isUserExists();
+  Future<void> removeUser();
+}
