+++ data/lib/repositories/settings_repository_impl.dart
@@ -1,11 +1,11 @@
-import 'package:data/providers/local/local_provider.dart';
+import 'package:data/providers/local/theme_data_provider.dart';
 import 'package:domain/domain.dart';
 
 class SettingsRepositoryImpl extends SettingsRepository{
-  final LocalProvider _localProvider;
+  final ThemeDataProvider _localProvider;
 
   SettingsRepositoryImpl({
-    required LocalProvider localProvider
+    required ThemeDataProvider localProvider
   }) : _localProvider = localProvider;
   
   @override
